requirements.txt
telethon
python-dotenv
loguru
fastapi
uvicorn
apscheduler
cryptg